Changing the url used for universal linking
Usng the existing url broke some apps which were using embedded UIWebView for oauth and inspecting the delegate callbacks to pull out the keys. We are now using a totally new url for our universal link, that will make sure only people intended to get bounced to native auth do so. This should be supported in Foursquare for iOS 9.6 in the next few days.